Remove  myobj
To just create a new, random instance is the the way to go here.
Even if the compiler accepted it the new object would be out of
context with no relation to the plugin etc.

Creating global, static objects here should be done with great
discretion, it is almost never the right solution.