Refactor wxEvtHandler handling
Problem with current code is that the link from the PreferenceDlg
to the plugin creating is broken. Partr of the drama is using
composition rather inheritance to make an wxEvtHandler available.

Retire the separate wxEvtHandler object and add it as a parent to
the plugin instead. This means all sorts of consequences, notably
how to connect GUI objects to events.